more of the same

+++ OpenGLTest1/OpenGLTest1/Meshes.h
@@ -1,6 +1,8 @@
-const unsigned int nTriangleCount = 2U;
-const GLfloat anMeshVertices[6][2] = //6 vertices, 2 components each
+//http://www.opengl-tutorial.org/beginners-tutorials/tutorial-7-model-loading/
+
+const size_t nTriangleCount = 2U;
+const double anMeshVertices[6][2] = //6 vertices, 2 components each
 {
-	{0.0f, 0.0f}, {0.1f, 0.0f}, {0.0f, 0.1f},
-	{0.0f, 0.1f}, {0.1f, 0.0f}, {0.1f, 0.1f},
+	{0.0, 0.0}, {0.1, 0.0}, {0.0, 0.1},
+	{0.0, 0.1}, {0.1, 0.0}, {0.1, 0.1},
 };
